backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.2
groq>=0.11.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
